Update to 0.40.3.
**Changelog:**

Hugo 0.40.3 fixes a possible .Content truncation issue introduced in 0.40.1. This should be very rare. It has been reported by only one user on a synthetic site. We have tested a number of big sites that does not show this problem with 0.40.2, but this is serious enough to warrant a patch release.